Replace Type Code with State/Strategy AND Replace Conditional with Polymorphism!
The final piece of code that stands out is the switch statement.  First, we see that there are
different types of movies.  We don't want to subclass movie directly, rather, since we notice
that the movie types are used only in the getCharge and getFrequentRenterPoints methods, we
can say that all movies regardless of type _have_ a PricingStrategy.

Then we simply create a PricingStrategy interface and an implementations of this interface
for each of the types.

The Groovy, and also the functional way, to implement the Strategy pattern is to pass closures
into the methods that need to perform calculations (think foldLeft function)
(see: http://groovy.codehaus.org/Strategy+Pattern); however, I wanted to at least have an
explicit interface rather than passing closures into the getCharge and
getFrequentRenterPoints methods.  In Groovy, I can define an interface and as its public static
fields define implementations of this interface as maps.  This may be an unconventional design
- one that doesn't quite fit the open/closed principle, but I think it is one interesting
way to Groovily implement this refactoring.